+++ WorkOrderDesk.Application/WorkOrders/DeleteWorkOrder/DeleteWorkOrderCommand.cs
@@ -1,0 +1,6 @@
+namespace WorkOrderDesk.Application.WorkOrders.DeleteWorkOrder;
+
+public sealed class DeleteWorkOrderCommand
+{
+    public Guid Id { get; init; }
+}